+++ package.json
@@ -44,7 +44,6 @@
     "path": "^0.12.7",
     "react": "^16.8.2",
     "react-dom": "^16.8.2",
-    "react-hot-loader": "^4.13.0",
     "react-rating-stars-component": "^2.2.0",
     "react-router-dom": "^5.0.1",
     "sass": "^1.27.1",
