--- src/client/components/Helper/Modal.js
@@ -1,23 +1,0 @@
-import React, { useEffect } from 'react';
-
-const Modal = () => {
-    useEffect(() => {
-        document.addEventListener('DOMContentLoaded', function () {
-            var elems = document.querySelectorAll('.modal');
-            var instances = M.Modal.init(elems, {});
-        });
-    }, []);
-    return (
-        <div id="modal1" className="modal">
-            <div className="modal-content">
-                <h4>Modal Header</h4>
-                <p>A bunch of text</p>
-            </div>
-            <div className="modal-footer">
-                <a href="#!" className="modal-close waves-effect waves-green btn-flat">Agree</a>
-            </div>
-        </div>
-    );
-}
-
-export default Modal;
