+++ src/client/components/Meals/MealReviews.js
@@ -5,6 +5,7 @@ import { useParams } from "react-router";
 import ReactStars from "react-rating-stars-component";
 import Moment from 'moment';
 import Pics from '../Helper/Pics';
+import { Link } from 'react-router-dom';
 
 const Mealreviews = (props) => {
     const params = useParams();
@@ -156,7 +157,6 @@ const Mealreviews = (props) => {
                     </div>
                 </div>
             </div>}
-
             <Footer />
         </div>
     );
